requirements.txt
fastapi~=0.121.1
pydantic~=2.12.4
pydantic_settings~=2.12.0
python-dotenv~=1.2.1
slowapi~=0.1.9
SQLAlchemy~=2.0.44
httpx~=0.28.0
uvicorn~=0.38.0
psycopg2-binary~=2.9.11
